onedrive_observatory/requirements.txt
pandas
numpy
plotly
pillow
statsmodels
